+++ test/Feature/CleanRegex/Match/stream/distinct/MatcherTest.php
@@ -25,7 +25,7 @@ class MatcherTest extends TestCase
         $stream = Pattern::of('Foo')
             ->match('Foo')
             ->stream()
-            ->flatMapAssoc(Functions::constant(['12', '12', 12, '12' => 13]));
+            ->toMap(Functions::constant(['12', '12', 12, '12' => 13]));
         // when
         $distinct = $stream->distinct()->all();
         // then